Created About page, linked CSS, and routed nav
Added CSS for IE viewport bug. Linked style.css properly so it shows up
in the app, not just HTML docs. Created the About page. Re-did the nav
section so it's in its own document and links back to the pages with
EJS. Nav section is now customizable for all pages with only making the
changes in one place.  Re-did sign up form so it has fields for names.